requirements.txt
# conda:
pytorch==2.2.2
accelerate==0.28.0
einops==0.7.0
matplotlib==3.7.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
scipy==1.12.0
tqdm==4.65.0
transformers==4.31.0
py-cpuinfo==8.0.0
pydantic==2.12.5
deepspeed==0.14.0
sentencepiece==0.2.0
peft==0.17.1
